+++ README.md
@@ -1,20 +1,13 @@
 # Prototype System
 
-A Prototype System to easily load dynamic content from files. 
+A Prototype System to easily load dynamic content from files.
 
-## Installation
+## Overview
 
-The easiest to install is to just add this project as a submodule in your project.
+## Installation Instructions
 
-Example usage in a Unity Project with git initialized:
+Follow the instructions in the [Unity documentation to install a package from a Git URL](https://docs.unity3d.com/6000.1/Documentation/Manual/upm-ui-giturl.html):
 
-1. Make sure your Unity Project is already in an initialized git project.
+## Workflows
 
-2. Open a Terminal window and type the following:
-
-```
-cd Path/of/UnityProject/Assets/Scripts
-git submodule add git@github.com:flaviakim/PrototypeSystem.git
-```
-
-3. Add the required Newtonsoft Json Package to your Unity Project as explained [here](https://github.com/applejag/Newtonsoft.Json-for-Unity/wiki/Install-official-via-UPM#installing-the-package-via-upm-window)
+## Samples
