--- Runtime/AssetSavingLoading/AssetSaverLoader.cs
@@ -1,9 +1,0 @@
-using UnityEngine;
-
-namespace AssetSavingLoading {
-    public static class AssetSaverLoader {
-        
-        public static string DefaultAssetPath { get; set; } = Application.streamingAssetsPath;
-        
-    }
-}
